refactor: user/answer/domain/entity의 클래스들을 global/entity로 이동
