+++ src/main/java/com/upstage/devup/global/entity/UserWrongAnswer.java
@@ -1,0 +1,34 @@
+package com.upstage.devup.global.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "wrong_answers")
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserWrongAnswer {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "question_id", nullable = false)
+    private Question question;
+
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+}
