+++ src/main/java/com/upstage/devup/user/answer/service/UserAnswerSaveService.java
@@ -7,9 +7,9 @@ import com.upstage.devup.global.entity.Question;
 import com.upstage.devup.question.service.QuestionService;
 import com.upstage.devup.user.answer.domain.dto.UserAnswerDetailDto;
 import com.upstage.devup.user.answer.domain.dto.UserAnswerSaveRequest;
-import com.upstage.devup.user.answer.domain.entity.UserAnswer;
-import com.upstage.devup.user.answer.domain.entity.UserAnswerStat;
-import com.upstage.devup.user.answer.domain.entity.UserWrongAnswer;
+import com.upstage.devup.global.entity.UserAnswer;
+import com.upstage.devup.global.entity.UserAnswerStat;
+import com.upstage.devup.global.entity.UserWrongAnswer;
 import com.upstage.devup.user.answer.repository.AnswerUserRepository;
 import com.upstage.devup.user.answer.repository.UserAnswerRepository;
 import com.upstage.devup.user.answer.repository.UserAnswerStatRepository;
